+++ vertx-web-client/src/test/java/io/vertx/ext/web/client/WebClientDatabindTest.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright 2019 Red Hat, Inc.
+ *
+ * Red Hat licenses this file to you under the Apache License, version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.vertx.ext.web.client;
+
+import io.vertx.core.buffer.Buffer;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.client.jackson.WineAndCheese;
+import io.vertx.ext.web.codec.BodyCodec;
+import org.junit.Test;
+
+/**
+ * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
+ */
+public class WebClientDatabindTest extends WebClientTestBase {
+
+  @Test
+  public void testResponseBodyAsJsonMapped() throws Exception {
+    JsonObject expected = new JsonObject().put("cheese", "Goat Cheese").put("wine", "Condrieu");
+    server.requestHandler(req -> req.response().end(expected.encode()));
+    startServer();
+    HttpRequest<Buffer> get = webClient.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/somepath");
+    get
+      .as(BodyCodec.json(WineAndCheese.class))
+      .send(onSuccess(resp -> {
+        assertEquals(200, resp.statusCode());
+        assertEquals(new WineAndCheese().setCheese("Goat Cheese").setWine("Condrieu"), resp.body());
+        testComplete();
+      }));
+    await();
+  }
+
+  @Test
+  public void testResponseUnknownContentTypeBodyAsJsonMapped() throws Exception {
+    JsonObject expected = new JsonObject().put("cheese", "Goat Cheese").put("wine", "Condrieu");
+    testResponseBody(expected.encode(), onSuccess(resp -> {
+      assertEquals(200, resp.statusCode());
+      assertEquals(new WineAndCheese().setCheese("Goat Cheese").setWine("Condrieu"), resp.bodyAsJson(WineAndCheese.class));
+      testComplete();
+    }));
+  }
+
+  @Test
+  public void testSendJsonPojoBody() throws Exception {
+    testSendBody(new WineAndCheese().setCheese("roquefort").setWine("Chateauneuf Du Pape"),
+      (contentType, buff) -> {
+        assertEquals("application/json", contentType);
+        assertEquals(new JsonObject().put("wine", "Chateauneuf Du Pape").put("cheese", "roquefort"), buff.toJsonObject());
+      });
+  }
+}
