+++ vertx-web-client/src/test/java/io/vertx/ext/web/client/WebClientTestBase.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2019 Red Hat, Inc.
+ *
+ * Red Hat licenses this file to you under the Apache License, version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.vertx.ext.web.client;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Handler;
+import io.vertx.core.VertxOptions;
+import io.vertx.core.buffer.Buffer;
+import io.vertx.core.dns.AddressResolverOptions;
+import io.vertx.core.http.HttpClientOptions;
+import io.vertx.core.http.HttpServerOptions;
+import io.vertx.core.http.HttpTestBase;
+import io.vertx.core.json.JsonObject;
+import org.junit.Rule;
+import org.junit.rules.TemporaryFolder;
+
+import java.util.function.BiConsumer;
+
+/**
+ * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
+ */
+public class WebClientTestBase extends HttpTestBase {
+
+  @Rule
+  public TemporaryFolder testFolder = new TemporaryFolder();
+
+  protected WebClient webClient;
+
+  @Override
+  protected VertxOptions getOptions() {
+    return super.getOptions().setAddressResolverOptions(new AddressResolverOptions().
+      setHostsValue(Buffer.buffer(
+        "127.0.0.1 somehost\n" +
+          "127.0.0.1 localhost")));
+  }
+
+  @Override
+  public void setUp() throws Exception {
+    super.setUp();
+    client = vertx.createHttpClient(new HttpClientOptions().setDefaultPort(8080).setDefaultHost("localhost"));
+    webClient = WebClient.wrap(client);
+    server.close();
+    server = vertx.createHttpServer(new HttpServerOptions().setPort(DEFAULT_HTTP_PORT).setHost(DEFAULT_HTTP_HOST));
+  }
+
+  protected void testResponseBody(String body, Handler<AsyncResult<HttpResponse<Buffer>>> checker) throws Exception {
+    server.requestHandler(req -> req.response().end(body));
+    startServer();
+    HttpRequest<Buffer> get = webClient.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/somepath");
+    get.send(checker);
+    await();
+  }
+
+  protected void testSendBody(Object body, BiConsumer<String, Buffer> checker) throws Exception {
+    waitFor(2);
+    server.requestHandler(req -> req.bodyHandler(buff -> {
+      checker.accept(req.getHeader("content-type"), buff);
+      complete();
+      req.response().end();
+    }));
+    startServer();
+    HttpRequest<Buffer> post = webClient.post(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/somepath");
+    if (body instanceof Buffer) {
+      post.sendBuffer((Buffer) body, onSuccess(resp -> complete()));
+    } else if (body instanceof JsonObject) {
+      post.sendJsonObject((JsonObject) body, onSuccess(resp -> complete()));
+    } else {
+      post.sendJson(body, onSuccess(resp -> complete()));
+    }
+    await();
+  }
+}
